Added usings to code examples and/or articles where needed.
